PR #71 add test for eql? and isomorphism?
These revealed a bug in subgraph? Subset relation between directed and
undirected graph can not be defined easily. Should directed edges be
considered as a special case of undirected edges? The sematics should be
defined rigidly in graph.clj. Meanwhile we do not compare directed with
undirected graph. They are always considered as unequal.