Generalized weight-aggregation functions instead of only `+`
Made some simple changes to use a function `waf`, short for
`weight-aggregation function`. Defaults to `+`, as the affected graph
functions previously implicitly used.

Affected functions include `dijkstra-traverse`, `dijkstra-span`,
`bellman-ford-transform`, and `johnson`.

The reason for this commit is that weights won't always be simply added
- they might be multiplied, subtracted, or divided. Suppose one were to
perform unit conversion, for instance, where to get from one unit to
another one would multiply by a constant (or equivalently, divides by
the inverse of the constant). To aggregate the conversions needed from
seconds to nanoseconds using `all-pairs-shortest-paths`, one would need
to not *add* 1/1000 to get to milliseconds, then another 1/1000 to get
to microseconds, then another to get to nanoseconds, but rather to
*multiply* them. There are certainly examples when weight-aggregation
functions might be more complex than simply `*` or `+`, but I will not
touch on them here.