Change loom.alg-generic/pre-traverse to do much less work
especially if the graph is dense with edges for a node that is
traversed, but the returned lazy sequence is not ever evaluated to get
to those nodes.

Also eliminate a stack-growing recursive call, using recur instead.
The previous version could grow the call stack nearly as deep as the
number of edges in the graph.